Terminar read_boolean()...
Si el primer carácter de la entrada es t o T retorna true.
Además:
* Corregir instrucción de comparación de <=.
* Sacar parámetros innecesarios de printf("true") y printf("false").

+++ src/ar/edu/unnoba/compilador/ast/expresiones/binarias/relaciones/MenorIgual.java
@@ -9,7 +9,7 @@ public class MenorIgual extends Relacion {
     }
 
     public String getInstruccionIR() {
-        return getIzquierda().getTipo() == Tipo.INTEGER ? "ole" : "sle";
+        return getIzquierda().getTipo() == Tipo.FLOAT ? "ole" : "sle";
     }
 
     @Override
